extract into rpc_send_request.go
